Coloquei no App.js
